Syscall: Re-order syscall9 to support cb on Linux
Note that this re-ordering is only really needed to support callbacks
with CGO_ENABLED=1. Most of the work already was fine for CGO_ENABLED=0
